requirements.txt
beautifulsoup4
curl_cffi
Flask
pandas
peewee
protobuf
python-dotenv
requests
websockets
yfinance
multitasking